changing the theme from light to dark and adding landing page
