+++ pages/conductor/deck/[id].js
@@ -335,6 +335,69 @@ const Edit = props => {
         }
     }
 
+    const roleSelectStyles = {
+        option: (provided, state) => {
+            let { side_id } = JSON.parse(state.value)
+            return {
+                ...provided,
+                backgroundColor:state.isFocused ? CreateSideColor(side_id.$oid).background  :  '#282828',
+                color:state.isFocused ? CreateSideColor(side_id.$oid).color  :  CreateSideColor(side_id.$oid).background,
+                cursor: 'pointer'
+            }
+        },
+        control: (provided, state) => {
+            return {
+                ...provided,
+                backgroundColor: '#3f3f3f',
+                border: 'none',
+                borderRadius: '8px'
+            }
+        },
+        multiValue: (provided, state) => ({
+            ...provided,
+            paddingRight: '4px',
+            backgroundColor:'#a688fa',
+            color:'#282828'
+            
+            
+        }),
+        menu: (provided, state) => ({
+            ...provided,
+            backgroundColor:'#282828'
+        }),
+
+    }
+
+    const cardSelectStyles = {
+        option: (provided, state) => {
+            return {
+                ...provided,
+                cursor: 'pointer',
+                backgroundColor: state.isFocused ?'#a688fa' :  '#282828',
+                color: state.isFocused ?'#121212' :  '#ffffffdd',
+            }
+        },
+        control: (provided, state) => {
+            return {
+                ...provided,
+                backgroundColor: '#282828',
+                border: 'none',
+                borderRadius: '8px'
+            }
+        },
+        multiValue: (provided, state) => ({
+            ...provided,
+            paddingRight: '4px',
+            backgroundColor:'#a688fa',
+            color:'#282828'
+
+        }),
+        menu: (provided, state) => ({
+            ...provided,
+            backgroundColor:'#282828'
+        }),
+    }
+
     return (
         <div className={styles.page}>
 
@@ -367,63 +430,12 @@ const Edit = props => {
 
                             <div className="row">
                                 <label htmlFor="roles">انتخاب نقش ها</label>
-                                <Select closeMenuOnSelect={false} menuPlacement="auto" minMenuHeight={300} placeholder={'انتخاب کنید'} styles={{
-                                    option: (provided, state) => {
-                                        let { side_id } = JSON.parse(state.value)
-                                        let color = {}
-                                        if (side_id.$oid === '630a35978c198e0d655c8adf') color = { color: '#549088' }
-                                        if (side_id.$oid === '630a359e8c198e0d655c8ae0') color = { color: '#cb5240' }
-                                        if (side_id.$oid === '630a35a38c198e0d655c8ae1') color = { color: '#ffde43' }
-                                        if (side_id.$oid === '630a3773696bb14037bdec19') color = { color: '#333333' }
-                                        if (side_id.$oid === '630a383d696bb14037bdec1a') color = { color: '#83bb70' }
-                                        if (side_id.$oid === '630a389a696bb14037bdec1b') color = { color: '#1d1a21' }
-                                        if (side_id.$oid === '630a3926696bb14037bdec1d') color = { color: '#333333' }
-                                        return {
-                                            ...provided,
-                                            ...color,
-                                            "&:hover": { ...CreateSideColor(side_id.$oid) },
-                                            cursor: 'pointer'
-                                        }
-                                    },
-                                    control: (provided, state) => {
-                                        return {
-                                            ...provided,
-                                            backgroundColor: '#E5E5E5',
-                                            border: 'none',
-                                            borderRadius: '8px'
-                                        }
-                                    },
-                                    multiValue: (provided, state) => ({
-                                        ...provided,
-                                        paddingRight: '4px',
-                                        border: '1px solid #CCC'
-                                    })
-                                }} value={defaultRoles} defaultValue={defaultRoles} options={roleOptions} id={"roles"} isRtl={true} isMulti={true} onChange={changeRoles} />
+                                <Select closeMenuOnSelect={false} menuPlacement="auto" minMenuHeight={300} placeholder={'انتخاب کنید'} styles={roleSelectStyles} value={defaultRoles} defaultValue={defaultRoles} options={roleOptions} id={"roles"} isRtl={true} isMulti={true} onChange={changeRoles} />
                             </div>
 
                             <div className="row">
                                 <label htmlFor="roles">انتخاب کارت های حرکت آخر</label>
-                                <Select closeMenuOnSelect={false} menuPlacement="auto" minMenuHeight={300} placeholder={'انتخاب کنید'} styles={{
-                                    option: (provided, state) => {
-                                        return {
-                                            ...provided,
-                                            cursor: 'pointer'
-                                        }
-                                    },
-                                    control: (provided, state) => {
-                                        return {
-                                            ...provided,
-                                            backgroundColor: '#E5E5E5',
-                                            border: 'none',
-                                            borderRadius: '8px'
-                                        }
-                                    },
-                                    multiValue: (provided, state) => ({
-                                        ...provided,
-                                        paddingRight: '4px',
-                                        border: '1px solid #CCC'
-                                    })
-                                }} value={defaultCards} defaultValue={defaultCards} options={cardOptions} id={"cards"} isRtl={true} isMulti={true} onChange={changeCards} />
+                                <Select closeMenuOnSelect={false} menuPlacement="auto" minMenuHeight={300} placeholder={'انتخاب کنید'} styles={cardSelectStyles} value={defaultCards} defaultValue={defaultCards} options={cardOptions} id={"cards"} isRtl={true} isMulti={true} onChange={changeCards} />
                             </div>
 
                             <div className="row">
