+++ pages/conductor/event/[id].js
@@ -109,17 +109,28 @@ const Create = props => {
      * @type {{control: (function(*, *): *&{border: string, backgroundColor: string, borderRadius: string}), multiValue: (function(*, *): *&{border: string, paddingRight: string}), option: (function(*, *): *&{cursor: string, "&:hover": {}})}}
      */
     const SelectStyles = {
+        option: (provided, state) => ({
+            ...provided,
+            backgroundColor: state.isFocused ? '#a688fa' : '#282828',
+            color: (state.isFocused) ? '#121212' : '#ffffffdd',
+            cursor: 'pointer'
+        }),
         control: (provided) => ({
             ...provided,
-            backgroundColor: '#E5E5E5',
+            backgroundColor: '#282828',
             border: 'none',
             borderRadius: '8px'
         }),
-        multiValue: (provided) => ({
+        singleValue: (provided, state) => ({
             ...provided,
-            paddingRight: '4px',
-            border: '1px solid #CCC'
-        })
+            color: '#ffffffdd',
+
+        }),
+        menu: (provided, state) => ({
+            ...provided,
+            backgroundColor: '#282828'
+        }),
+
     }
 
     /**
