+++ pages/old-login.js
@@ -4,7 +4,7 @@ import Image from "next/future/image";
 import React, { useEffect, useState } from "react";
 import Link from "next/link";
 import { toast, ToastContainer } from "react-toastify";
-import logo from "../public/logo.png";
+import logo from "../public/logo-white.png";
 import 'react-toastify/dist/ReactToastify.css';
 import { useRouter } from "next/router";
 import loginUser from "/utils/loginUser";
@@ -54,10 +54,10 @@ const Login = () => {
      */
     useEffect(() => {
         // if (hasCookie('token')) router.push(redirect).then()
-        if (globalUser && globalUser.isLoggedIn && globalUser.access_level === 1)
-            router.push('/profile')
+        if (globalUser && globalUser.isLoggedIn && globalUser.access_level === 2)
+            router.push('/home')
         else if (globalUser && globalUser.isLoggedIn)
-            router.push(redirect)
+            router.push('/profile')
     }, [redirect, router])
 
     /**
