+++ styles/components/nav.module.scss
@@ -3,8 +3,8 @@
   bottom: 0;
   right: 0;
   width: 100%;
-  border-top: 1px solid #e5e5e5;
-  background: #ffffff;
+  border-top: 1px solid #3f3f3f;
+  background-color: #282828;
   padding: 20px;
   z-index: 10;
