+++ styles/pages/auth/login-otp.module.scss
@@ -25,7 +25,7 @@
 }
 
 .card {
-  background: #ffffff;
+  background: var(--dark-bg-color);
   border-radius: 18px;
   box-shadow: 0px 0px 6px rgb(0, 0, 0, 0.5);
   min-width: 45vw;
