+++ styles/pages/conductor/deck/create.module.scss
@@ -17,6 +17,11 @@
         border-radius: 10px;
         padding: 16px;
         border: solid rgb(0, 0, 0, 0.16) 1px;
+        background-color: #282828;
+
+        input{
+        background-color: #282828;
+        }
       }
       label {
         font-size: 16px;
