+++ styles/pages/conductor/index.module.scss
@@ -12,8 +12,8 @@
       a {
         display: flex;
         width: 100%;
-        background: var(--primary-light-color);
-        color: var(--primary-color);
+        background: var(--primary-color);
+        color: var(--primary-light-color);
         font-size: 16px;
         text-align: center;
         padding: 12px;
