+++ styles/pages/event/info/player.module.scss
@@ -1,6 +1,6 @@
 .page {
   .card {
-    background: var(--primary-color);
+    background: #282828;
     padding: 20px;
     border-radius: 8px;
     position: relative;
@@ -14,8 +14,8 @@
       align-items: center;
       justify-content: center;
       font-size: 24px;
-      background: var(--text-icon-color);
-      color: var(--primary-color);
+      // background: var(--text-icon-color);
+      // color: var(--primary-color);
       padding: 0;
       width: 40px;
       height: 40px;
@@ -31,7 +31,7 @@
 
       img {
         border-radius: 100px;
-        border: 4px solid #E5E5E5;
+        border: 4px solid var(--primary-color);
         padding: 4px;
         margin-bottom: 10px;
       }
@@ -47,7 +47,8 @@
 
       .status {
         padding: 20px;
-        background: #FFFFFF;
+        background: var(--dark-bg-color);
+        border: solid 1px #3f3f3f;
         width: 100%;
         border-radius: 8px;
         display: flex;
@@ -56,7 +57,8 @@
           width: 100%;
 
           p {
-            font-weight: bold
+            font-weight: bold;
+            color: var(--primary-color);
           }
         }
       }
@@ -70,9 +72,10 @@
 
       tbody {
         tr {
-          border: 1px solid #E5E5E5;
+          border: solid 1px #3f3f3f;
           td {
             padding: 12px;
+            background-color: #282828;
 
             &:first-child {
               width: 100%;
