+++ styles/pages/explore.module.scss
@@ -17,7 +17,7 @@ $screen-xl-min: 1200px;
     padding: 14px;
   }
   .search {
-    border-bottom: 1px solid #e5e5e5;
+    border-bottom: 1px solid #3f3f3f;
     padding: 16px;
 
     label {
@@ -25,7 +25,6 @@ $screen-xl-min: 1200px;
       display: flex;
 
       input {
-        border: 1px solid #e5e5e5;
         border-radius: 40px;
         margin-bottom: 0;
       }
@@ -104,7 +103,7 @@ $screen-xl-min: 1200px;
           position: absolute;
           width: 100%;
           height: 100%;
-          background: rgb(0, 0, 0, 0.75);
+          background: rgba(0, 0, 0, 0.5);
         }
 
         img {
