+++ styles/pages/index.module.scss
@@ -1,193 +1,161 @@
-@import "styles/functions";
-
-// Small tablets and large smartphones (landscape view)
-$screen-sm-min: 576px;
-
-// Small tablets (portrait view)
-$screen-md-min: 768px;
-
-// Tablets and small desktops
-$screen-lg-min: 992px;
-
-// Large tablets and desktops
-$screen-xl-min: 1200px;
-
 .page {
-  .lastEvents {
-    // .swiper {
-    //   width: 90%;
-    //   padding-top: 50px;
-    //   padding-bottom: 50px;
-
-    //   @media (min-width: $screen-sm-min) {
-    //     width: 400px;
-    //   }
-    //   @media (min-width: $screen-md-min) {
-    //     width: 500px;
-    //   }
-    //   @media (min-width: $screen-lg-min) {
-    //     width: 600px;
-    //   }
-    //   @media (min-width: $screen-xl-min) {
-    //     width: 700px;
-    //   }
-
-    //   .swiper_slide {
-    //     aspect-ratio: 7/5;
-    //   }
-    //   a {
-    //     width: 100%;
-    //     height: 100%;
-    //     background-size: cover;
-    //     border-radius: 8px;
-    //     position: relative;
-    //     overflow: hidden;
-    //     display: flex;
-
-    //     &:before {
-    //       content: "";
-    //       position: absolute;
-    //       width: 100%;
-    //       height: 100%;
-    //       background: rgb(0, 0, 0);
-    //       background: -webkit-linear-gradient(
-    //         0deg,
-    //         rgba(0, 0, 0, 0.9) 0%,
-    //         rgba(0, 0, 0, 0) 100%
-    //       );
-    //       background: -moz-linear-gradient(
-    //         0deg,
-    //         rgba(0, 0, 0, 0.9) 0%,
-    //         rgba(0, 0, 0, 0) 100%
-    //       );
-    //       background: linear-gradient(
-    //         0deg,
-    //         rgba(0, 0, 0, 0.9) 0%,
-    //         rgba(0, 0, 0, 0) 100%
-    //       );
-    //     }
-
-    //     h3 {
-    //       position: absolute;
-    //       color: var(--text-icon-color);
-    //       bottom: 12px;
-    //       right: 12px;
-    //       padding-left: 24px;
-    //       max-height: 60px;
-    //       overflow: hidden;
-    //     }
-    //   }
-    // }
-
-    .event {
-      width: 100%;
-      margin: auto;
+    position: relative;
+  .header {
+    width: 80vw;
+    margin: auto;
+    /* margin-top: 6vh; */
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+    align-items: center;
+    padding: 0px;
+    h2 {
+      color: #ffffff;
+      font-family: "Outfit";
+      font-weight: 900;
+      font-size: 1.5rem;
+    }
+    nav {
       display: flex;
-      overflow: hidden;
-      aspect-ratio: 97/32;
+      flex-direction: row;
+      align-items: flex-start;
+      padding: 0px;
+      gap: 3vw;
+      a {
+        color: #ffffff;
+        font-family: "Outfit";
+        font-weight: 500;
+        font-size: 1rem;
+        text-decoration-line: none;
+        transition: 0.25s;
 
-      .item {
-        width: 100%;
-        height: 100%;
-        background-size: cover;
-        border-radius: 12px;
-        position: relative;
-        overflow: hidden;
-        display: flex;
-        align-items: flex-end;
-        flex-direction: column-reverse;
-        justify-content: end;
-
-        &::before {
-          content: "";
-          position: absolute;
-          width: 100%;
-          height: 100%;
-          background: rgb(0, 0, 0, 0.75);
-        }
-
-        img {
-          position: absolute;
-          display: block;
-          width: 100%;
-          height: 100%;
-          z-index: -1;
-        }
-
-        .data {
-          position: relative;
-          display: flex;
-          flex-direction: column;
-          padding: 12px 6px;
-          width: 100%;
-          gap: 5px;
-          .row {
-            display: flex;
-            flex-direction: row;
-            width: 90%;
-            margin: auto;
-            justify-content: space-between;
-            div {
-              display: flex;
-              flex-direction: column;
-            }
-          }
-          .event_title {
-            display: flex;
-            align-items: center;
-            gap: 3px;
-            svg {
-              fill: #cd0000;
-            }
-          }
-          h3 {
-            direction: rtl;
-            color: #eee;
-            font-size: 14px;
-            font-weight: normal;
-            white-space: nowrap;
-            overflow: hidden;
-            text-overflow: ellipsis;
-          }
-          span {
-            width: 100px;
-            padding-right: 15px;
-            color: #bbb;
-            font-size: 10px;
-            white-space: nowrap;
-            overflow: hidden;
-            text-overflow: ellipsis;
-
-            &.role {
-              display: inline-block;
-              font-size: 12px;
-              border-radius: 5px;
-              padding-right: 0;
-              height: 80%;
-              display: flex;
-              align-items: center;
-              justify-content: center;
-              align-self: center;
-            }
-          }
-        }
-        .btn {
-          position: relative;
-          background-color: transparent;
-          border: solid 2px;
-          color: #ddd;
-          font-size: 12px;
-          display: flex;
-          align-items: center;
-          padding: 0px 14px;
-          border-radius: 40px;
-          left: 12px;
-          top: 20px;
-          @media (max-width: 345px) {
-            top: 40px;
-          }
+        &:hover {
+          color: #ce643a;
+          transition: 0.25s;
         }
       }
     }
+  }
+
+  .svgnav {
+    display: none;
+  }
+  .search svg {
+    width: 18px;
+    height: 18px;
+    fill: #ffffff;
+    transition: 0.25s;
+  }
+  .search:hover svg {
+    fill: #ce643a;
+    transition: 0.25s;
+  }
+  .search svg:last-child {
+    position: relative;
+    width: 3px;
+    height: 3px;
+    right: 7px;
+  }
+
+
+
+  .hero {
+    width: 81.2vw;
+    height: 75vh;
+    margin: auto;
+    margin-top: 8vh;
+    display: flex;
+    justify-content: space-between;
+  }
+  .image {
+    width: 400px;
+    height: 400px;
+    background: url("/background.png");
+    background-size: cover;
+    transition: 0.25s;
+  }
+
+  .content {
+    // position: absolute;
+    width: 40vw;
+    height: 62vh;
+    bottom: 0;
+    h1 {
+      color: #ffffff;
+      font-family: "Outfit";
+      font-weight: 900;
+      font-size: 3.5rem;
+      line-height: 120%;
+    }
+    p {
+      color: #bdbdbd;
+      font-family: "Outfit";
+      font-weight: 400;
+      font-size: 1rem;
+      margin-top: 3vh;
+      line-height: 160%;
+    }
+  }
+
+  .frame {
+    margin-top: 7.3vh;
+    width: 100%;
+  }
+  .frame a {
+    display: inline-block;
+    text-decoration: none;
+    font-family: "Outfit";
+    font-weight: 600;
+    font-size: 1rem;
+    transition: 0.25s;
+  }
+  .hero-btn {
+    color: #ffffff;
+    border: 1px solid #ce643a;
+    border-radius: 0.75rem;
+    padding: 2.3vh 4vw;
+    background-color: #ce643a;
+    color: #ffffff;
+  }
+  .hero-btn:hover {
+    border: 1px solid #ce643a;
+    background-color: transparent;
+    color: #ce643a;
+  }
+  .btn {
+    border: 1px solid transparent;
+    margin-left: 1.17vw;
+    padding: 2.3vh 4.2vw;
+    background-color: transparent;
+    color: #ce643a;
+  }
+  .btn:hover {
+    color: white;
+  }
+  .arrow {
+    width: 32px;
+    height: 32px;
+    position: absolute;
+    bottom: 0;
+  }
+  .arrow svg:first-child {
+    position: absolute;
+    width: 2px;
+    height: 26px;
+    left: 12px;
+    bottom: 0;
+    fill: #ffffff;
+  }
+  .arrow svg:last-child {
+    position: absolute;
+    width: 20px;
+    height: 11px;
+    left: 3px;
+    bottom: 0;
+    fill: #ffffff;
+  }
+  .arrow:hover svg {
+    fill: #ce643a;
+    transition: 0.25s;
   }
 }
