+++ styles/pages/learn/index.module.scss
@@ -12,7 +12,7 @@ $screen-xl-min: 1200px;
 
 .page {
   .search {
-    border-bottom: 1px solid #e5e5e5;
+    border-bottom: 1px solid #3f3f3f;
     padding: 16px;
 
     label {
@@ -20,7 +20,6 @@ $screen-xl-min: 1200px;
       display: flex;
 
       input {
-        border: 1px solid #e5e5e5;
         border-radius: 40px;
         margin-bottom: 0;
       }
@@ -31,6 +30,7 @@ $screen-xl-min: 1200px;
         transform: translateY(-50%);
         font-size: 24px;
         left: 16px;
+        fill: var(--text-color);
       }
     }
   }
