+++ styles/pages/profile/index.module.scss
@@ -12,7 +12,7 @@
         width: 100%;
         text-align: center;
         padding: 12px 0;
-        border-bottom: 1px solid #e5e5e5;
+        border-bottom: 1px solid #3f3f3f;
         cursor: pointer;
 
         &.active {
@@ -62,7 +62,7 @@
                 position: absolute;
                 width: 100%;
                 height: 100%;
-                background: rgb(0, 0, 0, 0.75);
+                background: rgb(0, 0, 0, 0.5);
               }
 
               img {
